Fixed a lot of compiler warnings, there is still two left which i will leave to someone else. Added fps meter in the titlebar and also made so the window on startup spawn in the center of the screen.

+++ DV1573---UD1448/Gui/Gui.cpp
@@ -182,7 +182,7 @@ void Gui::update(float dt)
 	{
 		double x, y;
 		glfwGetCursorPos(glfwGetCurrentContext(), &x, &y);
-		m_context->injectMousePosition(x,y);
+		m_context->injectMousePosition(static_cast<float>(x), static_cast<float>(y));
 	}
 
 	if (Input::isAnyKeyPressed())
