+++ DV1573---UD1448/Mesh/Mesh.cpp
@@ -82,7 +82,7 @@ void Mesh::setUpBuffers()
 	glVertexAttribPointer(2, 2, GL_FLOAT, GL_FALSE, sizeof(Vertices), (void*)offsetof(Vertices, UV));
 	glBindVertexArray(0);
 
-	vertexBuffer.nrOfFaces = faces.size();
+	vertexBuffer.nrOfFaces = static_cast<int>(faces.size());
 }
